Fix update_dependencies code

+++ btsoot.py
@@ -187,11 +187,14 @@ def main():
 
 
 		elif sys.argv[1] == "update_dependencies":
-			print("This is not supported right now. Please delete them manualy and restart the program")
-			os.system("rm -rf compare")
-			os.system("git clone https://git.paukra.com/open-source/compare.git")
-			os.system("rm -rf datatransfer")
-			os.system("git clone https://git.paukra.com/open-source/datatransfer.git")
+			print("This requires an internet connection. ")
+			if input("Should i continue?") == "y":
+				os.system("rm -rf compare")
+				os.system("git clone https://git.paukra.com/open-source/compare.git")
+				os.system("rm -rf datatransfer")
+				os.system("git clone https://git.paukra.com/open-source/datatransfer.git")
+			else:
+				print(color.FAIL + "Abort." + color.ENDC)
 
 		else:
 			print(usage)
